% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_app}
\alias{build_app}
\title{build the app}
\usage{
build_app(name, ui_call, server_call, use_shiny_dashboard)
}
\arguments{
\item{name}{a character, the name of the module}

\item{ui_call}{a call, as returned by \code{\link[=build_ui_call]{build_ui_call()}}}

\item{server_call}{a call, as returned by \code{\link[=build_server_call]{build_server_call()}}}

\item{use_shiny_dashboard}{a logical, set to TRUE if the module uses \emph{shinydashboard} components}
}
\description{
build the app
}
