% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_params}
\alias{build_params}
\title{build the list of parameters to run_module}
\usage{
build_params(x)
}
\arguments{
\item{x}{the result of a \code{\link[=match.call]{match.call()}}}
}
\description{
build the list of parameters to run_module
}
