% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_server_call}
\alias{build_server_call}
\title{build the call for the server}
\usage{
build_server_call(name, dots)
}
\arguments{
\item{name}{a character, the name of the server function}

\item{dots}{a list of parameters, as returned by \code{\link[=build_params]{build_params()}}}
}
\description{
build the call for the server
}
