% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_ui_call}
\alias{build_ui_call}
\title{build the call for the UI}
\usage{
build_ui_call(name, dots)
}
\arguments{
\item{name}{a character, the name of the ui function}

\item{dots}{a list of parameters, as returned by \code{\link[=build_params]{build_params()}}}
}
\description{
build the call for the UI
}
