% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_ui_name}
\alias{find_ui_name}
\title{find the name of the UI function}
\usage{
find_ui_name(name)
}
\arguments{
\item{name}{a character, the name of the server function}
}
\description{
find the name of the UI function
}
