% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_module.R
\name{run_module}
\alias{run_module}
\title{run a module in standalone mode}
\usage{
run_module(module_name, ..., use_shiny_dashboard = FALSE)
}
\arguments{
\item{module_name}{a character, the name of the \strong{server} function}

\item{...}{additional named parameters passed to the module}

\item{use_shiny_dashboard}{a logical, set to TRUE if the module uses \emph{shinydashboard} components}
}
\description{
run a module in standalone mode
}
