% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyrunmodule-package.R
\docType{package}
\name{shinyrunmodule-package}
\alias{shinyrunmodule}
\alias{shinyrunmodule-package}
\title{shinyrunmodule: Run Shiny Modules as Standalone Apps}
\description{
Run a Shiny module as a standalone app for easier iteration.
}
\author{
\strong{Maintainer}: Pierre Formont \email{pierre.formont@gmail.com}

}
\keyword{internal}
